Add 404 page.

+++ hugo/deploy.sh
@@ -2,7 +2,9 @@
 
 set -Eeuo pipefail
 
-[[ -d public ]] && rm -r public
+if [[ "$#" -gt 0 && "$1" == "clean" ]]; then
+    [[ -d public ]] && rm -r public
+fi
 
 hugo
 aws s3 sync --profile soxley --delete public s3://stevenoxley-com-site
